+++ src/quail/StateData.hpp
@@ -23,7 +23,7 @@ public:
     void setBatteryVoltageState(float current); // sets the battery voltage state value
     void fireEmatch(uint8_t ematch_ch, bool updateEMs = true); // sets an ematch state to "fired"
     float getSensorState(uint8_t ch_id); // returns the corresponding state value for a given sensor
-    uint8_t getSolenoidState(); // returns the bit map of ematch states
+    uint16_t getSolenoidState(); // returns the bit map of solenoid states
     uint8_t getEmatchState(); // returns the bit map of ematch states
     float getBatteryCurrentState(); // returns the bit map of battery voltage state
     float getBatteryVoltageState(); // returns the bit map of battery current state
@@ -40,7 +40,7 @@ private:
     float batterycurrentstate; // float storing battery current measurement
     float batteryvoltagestate; // float storing battery voltage measurement
     static StaticJsonDocument<1024> stateJSON; // static JSON used to store sensor data for TX
-    uint8_t solenoidstate; // 8bit indicator of solenoid status (LSB is first solenoid channel, MSB is last channel)
+    uint16_t solenoidstate; // 16bit indicator of solenoid status (LSB is first solenoid channel, MSB is last channel)
     uint8_t ematchstate; // 8bit indicator of ematch status (LSB is first ematch channel, MSB is last channel)
     char currSequence[MAX_TEST_FILENAME_SIZE+1];
     static Mutex SVmutex; // mutex for solenoid valve state
